Fix insertion of spurious empty secondary sprites

+++ src/ppu/oam.rs
@@ -158,7 +158,7 @@ impl SecondaryOamPointer {
             None
         } else {
             Some(SecondaryOamPointer {
-                pointer: usize::from(value),
+                pointer: value,
                 end_reached: false,
             })
         }
